Function: Added the search functionality.
This functions computed first so we don't need to trigger. It purpose is
to find all the original lessons and the filter then to check if the are
included in the search input the user is making. So if it's true we
return the lessons similar to the search Query.