Functionality: changed the search, added ratings,
Added three functionality:
-search functionality.
first the changed the search, what i have done
is i have implemented the  loop through all the lessons and check if the
input search is matching with what i have in my lessons the return the
new lessons.

- Ratings:
added the ratings so that to know the lessons with high performance and
good attendance. i created a variable in my data rating, and the looped
through it using v-for in  and displayed it using an star emoji.

- Added sort by rating and spaces.

- In my form i added the required attribute in my forms, which specifies
  that a particular input field must be filled.